Refactoring
- extract the Weather Button to its own file
