+++ SwiftUI-Weather/SwiftUI-Weather/ContentView.swift
@@ -117,18 +117,3 @@ struct MainWeatherStatusView: View {
         .padding(.bottom, 40)
     }
 }
-
-struct WeatherButton: View {
-    var title: String
-    var textColor: Color
-    var backgroundColor: Color
-    
-    var body: some View {
-        Text(title)
-            .frame(width: 280, height: 50)
-            .background(Color.white)
-            .foregroundColor(.blue)
-            .font(.system(size: 20, weight: .bold, design: .default))
-            .cornerRadius(10)
-    }
-}
