+++ SwiftUI-Weather/SwiftUI-Weather/WeatherButton.swift
@@ -1,0 +1,23 @@
+//
+//  WeatherButton.swift
+//  SwiftUI-Weather
+//
+//  Created by Norbert Agoston on 12.02.2024.
+//
+
+import SwiftUI
+
+struct WeatherButton: View {
+    var title: String
+    var textColor: Color
+    var backgroundColor: Color
+    
+    var body: some View {
+        Text(title)
+            .frame(width: 280, height: 50)
+            .background(Color.white)
+            .foregroundColor(.blue)
+            .font(.system(size: 20, weight: .bold, design: .default))
+            .cornerRadius(10)
+    }
+}
